rearrange fixtures
