+++ ml_project/tests/pipelines/test_prediction.py
@@ -1,6 +1,11 @@
 import os
-from typing import Dict
+from typing import Dict, List
 
+from py._path.local import LocalPath
+import pytest
+import pandas as pd
+
+from ..helpers import generate_fake_df
 from src.parameters import (
     PathParams,
     PreprocessingParams,
@@ -11,6 +16,44 @@ from src.pipelines import (
     run_training_pipeline,
     run_prediction_pipeline,
 )
+
+
+@pytest.fixture(scope="function")
+def fake_test_dataset(
+    feature_columns: List,
+    target_col: str,
+) -> pd.DataFrame:
+    df = generate_fake_df(
+        feature_columns,
+        size=1111,
+    )
+    return df
+
+
+@pytest.fixture(scope="function")
+def fake_test_dataset_path(
+    tmpdir: LocalPath,
+    fake_test_dataset: pd.DataFrame,
+) -> LocalPath:
+    path = tmpdir.join("fake_test_data.csv")
+    fake_test_dataset.to_csv(path, index=False)
+    return path
+
+
+@pytest.fixture(scope='function')
+def prediction_params(
+    fake_test_dataset_path: LocalPath,
+    model_path: LocalPath,
+    transformer_path: LocalPath,
+    predictions_path: LocalPath,
+) -> PredictionParams:
+    params = PredictionParams(
+        dataset_path=fake_test_dataset_path,
+        model_path=model_path,
+        transformer_path=transformer_path,
+        predictions_path=predictions_path,
+    )
+    return params
 
 
 def test_prediction(
